Complete Xcode and deal with wxWidget version 3.1.4. Remove all warnings and replace deprecated functions. Dark mode icons corrected to have no white background. Dark mode color set to macOS example. New setting to remove menu separators, as menu is very long on macOS. Re-open/unlock menu dialog deals with read-only flag.

+++ src/core/ItemData.cpp
@@ -176,7 +176,7 @@ int CItemData::Read(PWSfile *in)
           trashMemory(utf8, utf8Len * sizeof(utf8[0]));
           delete[] utf8;
         }
-        return -numread;
+        return static_cast<int>(-numread);
       } else if (type != END) { // unknown field
         SetUnknownField(type, utf8Len, utf8);
       }
